+++ src/game.c
@@ -4,7 +4,7 @@
 
 #include <stddef.h>
 
-void game_init(struct Game *game, size_t width, size_t height, size_t mines) {
+void game_init(struct Game *game, int width, int height, int mines) {
     game->state = ALIVE;
     minefield_init(&game->minefield, width, height, mines);
 }
@@ -13,13 +13,13 @@ void game_cleanup(struct Game *game) {
     minefield_cleanup(&game->minefield);
 }
 
-void game_click_tile(struct Game *game, size_t x, size_t y) {
+void game_click_tile(struct Game *game, int x, int y) {
     game_undo_store(game);
     bool still_alive = minefield_reveal_tile(&game->minefield, x, y); /* false if dead from clicking a mine */
     if (!still_alive) {
         game->state = DEAD;
         /* reveal all the mines */
-        size_t x, y;
+        int x, y;
         for (x = 0; x < game->minefield.width; x++) {
             for (y = 0; y < game->minefield.height; y++) {
                 if (minefield_get_tile(&game->minefield, x, y)->mine) {
@@ -29,7 +29,7 @@ void game_click_tile(struct Game *game, size_t x, size_t y) {
         }
     } else if (minefield_check_victory(&game->minefield)) {
         game->state = VICTORY;
-        size_t x, y;
+        int x, y;
         for (x = 0; x < game->minefield.width; x++) {
             for (y = 0; y < game->minefield.height; y++) {
                 minefield_get_tile(&game->minefield, x, y)->visible = true;
