+++ src/minefield.c
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdlib.h>
 
-bool minefield_init(struct Minefield *minefield, size_t width, size_t height, size_t mines) {
+bool minefield_init(struct Minefield *minefield, int width, int height, int mines) {
     minefield->width = width;
     minefield->height = height;
     minefield->mines = mines;
@@ -30,17 +30,17 @@ void minefield_cleanup(struct Minefield *minefield) {
 }
 
 /* set a tile as a mine and increment surrounding */
-static void minefield_set_mine(struct Minefield *minefield, size_t x, size_t y) {
+static void minefield_set_mine(struct Minefield *minefield, int x, int y) {
     struct Tile *tile = minefield_get_tile(minefield, x, y);
     tile->mine = true;
 
-    size_t x_start = x > 0 ? x - 1 : 0;
-    size_t y_start = y > 0 ? y - 1 : 0;
+    int x_start = x > 0 ? x - 1 : 0;
+    int y_start = y > 0 ? y - 1 : 0;
     /* TODO: this is kinda ugly */
-    size_t x_end = x < minefield->width - 1 ? x + 1 : x;
-    size_t y_end = y < minefield->height - 1 ? y + 1 : y;
+    int x_end = x < minefield->width - 1 ? x + 1 : x;
+    int y_end = y < minefield->height - 1 ? y + 1 : y;
 
-    size_t x1, y1;
+    int x1, y1;
     for (x1 = x_start; x1 <= x_end; x1++) {
         for (y1 = y_start; y1 <= y_end; y1++) {
             minefield_get_tile(minefield, x1, y1)->surrounding++;
@@ -49,11 +49,11 @@ static void minefield_set_mine(struct Minefield *minefield, size_t x, size_t y) 
 }
 void minefield_populate(struct Minefield *minefield) {
     /* randomly spread mines */
-    size_t i;
+    int i;
     for (i = 0; i < minefield->mines;) {
         /* non inclusive; don't worry, i didn't forget about starting at 0 */
-        size_t x = (rand() % (minefield->width));
-        size_t y = (rand() % (minefield->height));
+        int x = (rand() % (minefield->width));
+        int y = (rand() % (minefield->height));
 
         /* TODO: maybe calculate using distance formula */
         /* don't generate mines in a 3x3 centered on the cursor */
@@ -72,14 +72,14 @@ void minefield_populate(struct Minefield *minefield) {
     }
 }
 
-struct Tile *minefield_get_tile(struct Minefield *minefield, size_t x, size_t y) {
+struct Tile *minefield_get_tile(struct Minefield *minefield, int x, int y) {
     /* tile array is treated as a sequential list of rows, each row containing `minefield.cols` elements */
-    size_t offset = y * minefield->width;
+    int offset = y * minefield->width;
     return &minefield->tiles[offset + x];
 }
 
 /* output: bool - false if the clicked tile was a mine, true otherwise */
-bool minefield_reveal_tile(struct Minefield *minefield, size_t x, size_t y) {
+bool minefield_reveal_tile(struct Minefield *minefield, int x, int y) {
     struct Tile *tile = minefield_get_tile(minefield, x, y);
     assert(!tile->flagged);
     bool start_visible = tile->visible;
@@ -91,14 +91,14 @@ bool minefield_reveal_tile(struct Minefield *minefield, size_t x, size_t y) {
         return true;
     }
 
-    size_t x_start = x > 0 ? x - 1 : 0;
-    size_t y_start = y > 0 ? y - 1 : 0;
+    int x_start = x > 0 ? x - 1 : 0;
+    int y_start = y > 0 ? y - 1 : 0;
     /* TODO: this is kinda ugly */
-    size_t x_end = x < minefield->width - 1 ? x + 1 : x;
-    size_t y_end = y < minefield->height - 1 ? y + 1 : y;
+    int x_end = x < minefield->width - 1 ? x + 1 : x;
+    int y_end = y < minefield->height - 1 ? y + 1 : y;
     bool no_mines = true;
 
-    size_t x1, y1;
+    int x1, y1;
     for (x1 = x_start; x1 <= x_end; x1++) {
         for (y1 = y_start; y1 <= y_end; y1++) {
             struct Tile *surtile = minefield_get_tile(minefield, x1, y1);
@@ -110,10 +110,10 @@ bool minefield_reveal_tile(struct Minefield *minefield, size_t x, size_t y) {
     return no_mines;
 }
 
-size_t minefield_count_surrounding_mines(struct Minefield *minefield, size_t x, size_t y) {
-    size_t surrounding = 0;
+int minefield_count_surrounding_mines(struct Minefield *minefield, int x, int y) {
+    int surrounding = 0;
 
-    size_t x1, y1;
+    int x1, y1;
     for (x1 = x - 1; x1 <= x + 1; x1++) {
         for (y1 = y - 1; y1 <= y + 1; y1++) {
             if ((x1 >= 0 && y1 >= 0) && (x1 < minefield->width && y1 < minefield->height)) { /* make sure we are in bounds */
@@ -127,10 +127,10 @@ size_t minefield_count_surrounding_mines(struct Minefield *minefield, size_t x, 
     return surrounding;
 }
 
-size_t minefield_count_surrounding_flags(struct Minefield *minefield, size_t x, size_t y) {
-    size_t surrounding = 0;
+int minefield_count_surrounding_flags(struct Minefield *minefield, int x, int y) {
+    int surrounding = 0;
 
-    size_t x1, y1;
+    int x1, y1;
     for (x1 = x - 1; x1 <= x + 1; x1++) {
         for (y1 = y - 1; y1 <= y + 1; y1++) {
             if ((x1 >= 0 && y1 >= 0) && (x1 < minefield->width && y1 < minefield->height)) { /* make sure we are in bounds */
@@ -147,8 +147,8 @@ size_t minefield_count_surrounding_flags(struct Minefield *minefield, size_t x, 
 bool minefield_check_victory(struct Minefield *minefield) {
     /* TODO: count up the hidden tiles as they are revealed so they
      * don't have to be recounted every time this function runs */
-    size_t hidden = 0;
-    size_t x, y;
+    int hidden = 0;
+    int x, y;
     for (x = 0; x < minefield->width; x++) {
         for (y = 0; y < minefield->height; y++) {
             if (!minefield_get_tile(minefield, x, y)->visible)
